+++ src/classes/js/TSLinterErrorFactory.ts
@@ -1,12 +1,13 @@
-import { JSLinter } from "./parts/abstraction/JSLinter";
-import { UnusedMemberLinter } from "./parts/UnusedMemberLinter";
-import { WrongFilePathLinter } from "./parts/WrongFilePathLinter";
 import { AnyCustomTSClass, TextDocument, UI5TSParser } from "ui5plugin-parser";
 import { IError, Linter } from "../Linter";
-import { UnusedClassLinter } from "./parts/UnusedClassLinter";
-import { WrongClassNameLinter } from "./parts/WrongClassNameLinter";
-import { WrongNamespaceLinter } from "./parts/WrongNamespaceLinter";
+import { EventTypeLinter } from "./parts/EventTypeLinter";
 import { PublicMemberLinter } from "./parts/PublicMemberLinter";
+import { UnusedClassLinter } from "./parts/UnusedClassLinter";
+import { UnusedMemberLinter } from "./parts/UnusedMemberLinter";
+import { WrongClassNameLinter } from "./parts/WrongClassNameLinter";
+import { WrongFilePathLinter } from "./parts/WrongFilePathLinter";
+import { WrongNamespaceLinter } from "./parts/WrongNamespaceLinter";
+import { JSLinter } from "./parts/abstraction/JSLinter";
 
 export class TSLinterErrorFactory extends Linter<UI5TSParser, AnyCustomTSClass> {
 	timePerchar = 0;
@@ -17,7 +18,8 @@ export class TSLinterErrorFactory extends Linter<UI5TSParser, AnyCustomTSClass> 
 			new UnusedClassLinter(this._parser, this._configHandler),
 			new WrongClassNameLinter(this._parser, this._configHandler),
 			new WrongNamespaceLinter(this._parser, this._configHandler),
-			new PublicMemberLinter(this._parser, this._configHandler)
+			new PublicMemberLinter(this._parser, this._configHandler),
+			new EventTypeLinter(this._parser, this._configHandler)
 		];
 
 		const errors = linters.flatMap(linter => linter.getLintingErrors(document));